+++ src/background-plugin.ts
@@ -1,7 +1,7 @@
 /*eslint-disable no-empty*/
 import { CallContext, RESET, IMPORT_STATE, EXPORT_STATE, STORE, GET_BLOCK } from './call-context.ts';
 import { MemoryOptions, PluginOutput, SAB_BASE_OFFSET, SharedArrayBufferSection, type InternalConfig } from './interfaces.ts';
-import { withTimeout } from './utils.ts';
+import { readBodyUpTo, withTimeout } from './utils.ts';
 import { WORKER_URL } from './worker-url.ts';
 import { Worker } from 'node:worker_threads';
 import { CAPABILITIES } from './polyfills/deno-capabilities.ts';
@@ -37,7 +37,7 @@ const AtomicsWaitAsync =
   })();
 
 class BackgroundPlugin {
-  worker: Worker;
+  worker?: Worker;
   sharedData: SharedArrayBuffer;
   sharedDataView: DataView;
   hostFlag: Int32Array;
@@ -58,7 +58,7 @@ class BackgroundPlugin {
     this.opts = opts;
     this.#context = context;
 
-    // to shut up the type checker
+    // make the type checker happy, even though we're going to overwrite it
     this.worker = new Worker(WORKER_URL);
   }
 
@@ -71,13 +71,10 @@ class BackgroundPlugin {
     // Create a new worker
     this.worker = new Worker(WORKER_URL);
 
-    // Reinitialize the worker
     await this.initializeWorker();
 
-    // Reset the context
     this.#context[RESET]();
 
-    // Reset other necessary state
     this.#request = null;
     this.hostFlag[0] = SAB_BASE_OFFSET;
 
@@ -533,37 +530,6 @@ class HttpContext {
 
     return result;
   }
-}
-
-async function readBodyUpTo(response: Response, maxBytes: number): Promise<Uint8Array> {
-  const reader = response.body?.getReader();
-  if (!reader) {
-    return new Uint8Array(0);
-  }
-
-  let receivedLength = 0;
-  const chunks = [];
-
-  while (receivedLength < maxBytes) {
-    const { done, value } = await reader.read();
-    if (done) {
-      break;
-    }
-    chunks.push(value);
-    receivedLength += value.length;
-    if (receivedLength >= maxBytes) {
-      throw new Error(`Response body exceeded ${maxBytes} bytes`);
-    }
-  }
-
-  const limitedResponseBody = new Uint8Array(receivedLength);
-  let position = 0;
-  for (const chunk of chunks) {
-    limitedResponseBody.set(chunk, position);
-    position += chunk.length;
-  }
-
-  return limitedResponseBody;
 }
 
 export async function createBackgroundPlugin(
